PhpStan updates (#12)
- closes #11

+++ src/app/code/core/Mage/Admin/Model/Resource/User.php
@@ -293,7 +293,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     {
         $dbh = $this->_getWriteAdapter();
         $aRoles = $this->hasAssigned2Role($user);
-        if (count($aRoles)) {
+        if (is_array($aRoles) && count($aRoles)) {
             foreach ($aRoles as $data) {
                 $dbh->delete(
                     $this->getTable('admin/role'),
